requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tensorflow==2.16.2
pillow==10.4.0
numpy==1.26.4
pydicom==2.4.4
python-multipart==0.0.9
jinja2==3.1.4
